[Android] Move the shadow onto the tab bar for Android
On Android everything is `overflow: hidden` so we need to render the shadow within the bounds of the TabBar. We could add a transparent wrapper view around the tab bar so the shadow appears to sit outside of it, but I don't think it's currently worth it till we need it.