Allow child views to be updated as they are deselected
Currently it's impossible to pass selected down to child views
because the static container blocks their updates unless selected,
so they never know when selected changes to false.

The change here allows the child to get one final update as it is
deselected, then continues to block updates until it's selected again.