renderIcon is a weird mix of optional (TabNavigatorIcon has a default value, TabNavigator constantly checks item.props.renderIcon) and non-optional (marked as func.isRequired). On top of that, the default value of renderIcon returns a <View />, which does not accept the defaultSelectedIcon stylesheet's tintColor (which is only allowed on <Image /> elements.)
This correctly makes renderIcon optional again in all senses of the word, and ensures the code works without warnings in such cases. (I see some references on the issue tracker about this problem.)

This isn't important in shipping code (where you'll most likely have icons), but is important when first setting it up without icons and seeing scary looking warnings.
The renderIcon is optional in many ways (has a default value that returns a